// old/frontend-react/src/generated/graphql.tsx
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
};
// ...
export type MutationUpdatePlayerIdArgs = {
  discordId: Scalars["String"];
  playerId: Scalars["String"];
};